Split table component from txHistory component
